pyproject.toml
[tool.poetry]
name = "mes-printing-server"
version = "0.1.0"
description = "printing server of the MES system"
authors = ["Jicheng Zhi <jzhi0004@student.monash.edu>"]
license = "MIT"
readme = "README.md"
packages = [
    { include = "db", from = "src" },
    { include = "printer", from = "src" },
    { include = "service", from = "src" },
    { include = "scheduler", from = "src" },
    { include = "worker", from = "src" },
    { include = "app", from = "src" },
]

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.5.3"
aiohttp = "^3.9.1"
SQLAlchemy = "^2.0.23"
asyncua = "^1.0.6"
asyncpg = "^0.29.0"
python-dotenv = "^1.0.0"
sqlmodel = "^0.0.14"
fastapi = "^0.108.0"
uvicorn = "^0.25.0"
python-multipart = "^0.0.6"
aiofiles = "^23.2.1"
pydantic-settings = "^2.1.0"
python-rapidjson = "^1.14"
opcuax = { git = "https://github.com/monash-automation/opcuax.git" }
mes_opcua_server = { git = "https://github.com/monash-automation/mes-opcua-server.git" }
httpx = "^0.27.0"

[tool.poetry.group.dev.dependencies]
black = "^23.12.1"
pytest = "^7.4.3"
pytest-asyncio = "^0.23.2"
async-timeout = "*"
isort = "^5.13.2"
pre-commit = "^3.6.0"
mypy = "^1.8.0"
types-aiofiles = "^23.2.0.20240106"
pytest-aiohttp = "^1.0.5"
ruff = "^0.2.2"

[tool.mypy]
python_version = "3.11"
files = ["src/"]
plugins = ["pydantic.mypy"]
show_error_codes = true
pretty = true
follow_imports = "silent"
warn_redundant_casts = true
warn_unused_ignores = true
disallow_any_generics = true
check_untyped_defs = true
no_implicit_reexport = true

[[tool.mypy.overrides]]
module = ["asyncua", "rapidjson"]
ignore_missing_imports = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.poetry.scripts]
setting = "setting:display"
server = "app:run"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
